reorganize cylinder
