+++ Makefile
@@ -25,7 +25,9 @@ FILES =		destructor.c			\
 		mlx/mlx.c			\
 		objects/ambiant.c		\
 		objects/camera.c		\
-		objects/cylinder.c		\
+		objects/cylinder/caps.c		\
+		objects/cylinder/cylinder.c	\
+		objects/cylinder/side.c		\
 		objects/factory.c		\
 		objects/light.c			\
 		objects/plane.c			\
