+++ maps/mini.rt
@@ -3,5 +3,5 @@ C -50,10,20 0,0,1 70
 pl 0,0,0 0,1,0 0,255,0
 sp -60,7,50 10 100,24,0
 sp -30,5,50 10 255,255,0
-cy -45.0,10,50 0.5,0.5,1 14.2 21.42 255,0,255
+cy -45.0,7,50 0.5,0.5,1 14.2 21.42 255,0,255
 cy -65.0,25,50 1,1,0 5 12.42 100,100,0
